Mongo ReplicaSet creator.
- Adds cygwin support.
- Factors out common/duplicated code.
- Hardens parameter validation.

+++ mongo/replicaset.bash
@@ -19,12 +19,18 @@ set -e # this makes bash exit as soon as any of the
 
 #globals
 data_root=
-num_of_nodes=
+num_nodes=
+rset_file=
+start_file=
+stop_file=
 sleep_secs=7
 rsetid=rs_`date +%s`
-rset_file=${rsetid}.js
-start_file=${rsetid}.start
-stop_file=${rsetid}.stop
+cygwin=false
+cygregex="^cygwin.*$"
+shopt -s nocasematch
+if [[ `uname -s` =~ $cygregex ]]; then
+    cygwin=true
+fi
 
 #echoes the usage info
 function usage() {
@@ -49,29 +55,40 @@ function trim_trailing_slash() {
     fi
 }
 
-#grabs the number of shards and the number
-#of nodes in the shard replica set from the
-#command line
+#grabs the number of nodes and the data root directory
+#from the command line
 function process_args() {
     if [ $# -ne 2 ]; then
-        echo "ERROR: i need two arguments to run."
+        echo "ERROR: i need 2 arguments to run. i got $# instead."
         usage
         exit 1
-    else
-        num_of_nodes=$1
-        if [ ${num_of_nodes} -lt 2 ] || [ ${num_of_nodes} -gt 7 ]; then
-            echo "ERROR: the number of nodes per shard replica set '${num_of_nodes}' is outside of the valid range."
-            usage
-            exit 1
-        fi
-        data_root=`trim_trailing_slash $2`
-        echo "will create a replica set with '${num_of_nodes}' nodes (an arbiter will be added if the"
-        echo "nodes are even). the data will be hosted under the '${data_root}/${rsetid}' directory."
     fi
+
+    num_nodes=$1
+    if [[ ${num_nodes} != [0-9]* ]] || [ ${num_nodes} -lt 2 ] || [ ${num_nodes} -gt 7 ]; then
+        echo "ERROR: the number of nodes for the replica set '${num_nodes}' is outside of the valid range."
+        usage
+        exit 1
+    fi
+
+    data_root=`readlink -f $2`
+    data_root=`trim_trailing_slash ${data_root}`
+    if [[ ! -d ${data_root} ]]; then
+        echo "ERROR: the second argument (${data_root}) must be a directory."
+        usage
+        exit 1
+    fi
+
+    echo "will create a replica set with '${num_nodes}' nodes (an arbiter will be added if the"
+    echo "nodes are even). the data will be hosted under the '${data_root}/${rsetid}' directory."
+    echo "the control files will be hosted under the '${data_root}' directory."
+
+    rset_file=${data_root}/${rsetid}.js
+    start_file=${data_root}/${rsetid}.start
+    stop_file=${data_root}/${rsetid}.stop
 }
 
-#generates the mongo replica set config file for
-#the shard at the globally set shard_num
+#generates the mongo replica set config file
 function generate_rset_config() {
     if [ -e ${rset_file} ]; then
         echo "WARNING: the mongo replica set configuration file '${rset_file}' exists and will be overwritten."
@@ -81,25 +98,60 @@ function generate_rset_config() {
     fi
 
     local last_node_idx=-1
-    let "last_node_idx = ${num_of_nodes} - 1"
+    let "last_node_idx = ${num_nodes} - 1"
 
     echo "config = { _id: \"${rsetid}\", members: [" > ${rset_file}
-    for ((node = 0; node < ${num_of_nodes}; node++))
+    for ((node = 0; node < ${num_nodes}; node++))
     do
         node_cfg="{_id: ${node}, host: \"127.0.0.1:5000${node}\"}"
-        if [ ${node} -lt ${last_node_idx} ] || [ $(( ${num_of_nodes} % 2 )) -eq 0 ]; then
+        if [ ${node} -lt ${last_node_idx} ] || [ $(( ${num_nodes} % 2 )) -eq 0 ]; then
             node_cfg="${node_cfg},"
         fi
         echo ${node_cfg} >> ${rset_file}
     done
-    if [ $(( ${num_of_nodes} % 2 )) -eq 0 ]; then
-        echo "{_id: ${node}, host: \"127.0.0.1:5000${num_of_nodes}\", arbiterOnly: true}" >> ${rset_file}
+    if [ $(( ${num_nodes} % 2 )) -eq 0 ]; then
+        echo "{_id: ${node}, host: \"127.0.0.1:5000${num_nodes}\", arbiterOnly: true}" >> ${rset_file}
     fi
     echo "]}" >> ${rset_file}
 
     echo "rs.initiate(config)" >> ${rset_file}
 }
 
+#records the mongod start command in the start control file.
+function cmd_mongod() {
+    if [ $# -ne 2 ]; then
+        echo "FATAL: i need two arguments to start a mongod (the node number and name). i got $# arguments instead."
+        exit 1
+    fi
+
+    node_num=$1
+    if [[ ${node_num} != [0-9] ]]; then
+        echo "FATAL: the node number ('${node_num}') is outside of the valid range ('[0-9]')."
+        exit 1
+    fi
+
+    node_name=$2
+    if [[ ${node_name} =~ ^.*\ +.*$ ]]; then
+        echo "FATAL: the node name ('${node_name}') must not contain white space."
+        exit 1
+    fi
+
+    data_dir=${data_root}/${rsetid}/${node_name}
+    log_file=${data_dir}/mongod.log
+    outerr_file=${data_dir}/outerr.log
+    port=5000${node_num}
+    mkdir -p ${data_dir}
+    cmd="mongod --fork --logpath ${log_file} --port ${port} --replSet ${rsetid} --dbpath ${data_dir}"
+    if ${cygwin}; then
+        log_file='`cygpath -w '${log_file}'`'
+        data_dir='`cygpath -w '${data_dir}'`'
+        outerr_file='`cygpath -w '${outerr_file}'`'
+        cmd="mongod --logpath ${log_file} --port ${port} --replSet ${rsetid} --dbpath ${data_dir} > ${outerr_file} 2>&1 &"
+    fi
+    echo ${cmd} >> ${start_file}
+}
+
+# get and validate the arguments
 process_args "$@"
 
 #create the start file
@@ -107,38 +159,38 @@ touch ${start_file}
 chmod 755 ${start_file}
 echo "#!/usr/bin/env bash" >> ${start_file}
 
-#start the replica set for each shard
-for ((node = 0; node < ${num_of_nodes}; node++))
+#start mongo for each node in the set
+for ((node = 0; node < ${num_nodes}; node++))
 do
-    data_dir=${data_root}/${rsetid}/node${node}
-    log_file=${data_dir}/mongod.log
-    port=5000${node}
-    mkdir -p ${data_dir}
-    cmd="mongod --fork --logpath ${log_file} --port ${port} --replSet ${rsetid} --dbpath ${data_dir}"
-    echo ${cmd} >> ${start_file}
-    ${cmd}
+    cmd_mongod "${node}" "node${node}"
 done
 
 #spawn an arbiter only mongod process if we otherwise would
 #end up with an even number of nodes
-if [ $(( ${num_of_nodes} % 2 )) -eq 0 ]; then
-    data_dir=${data_root}/${rsetid}/arbiter
-    log_file=${data_dir}/mongod.log
-    port=5000${num_of_nodes}
-    mkdir -p ${data_dir}
-    cmd="mongod --fork --logpath ${log_file} --port ${port} --replSet ${rsetid} --dbpath ${data_dir}"
-    echo ${cmd} >> ${start_file}
-    ${cmd}
+if [ $(( ${num_nodes} % 2 )) -eq 0 ]; then
+    cmd_mongod "${num_nodes}" "arbiter"
 fi
 
 #create the stop file
 touch ${stop_file}
 chmod 755 ${stop_file}
 echo "#!/usr/bin/env bash" >> ${stop_file}
-echo "ps ax | grep mongod | grep ${rsetid} | grep -v grep | sed 's/^[ \t]*//' | cut -d' ' -f1 | xargs kill -2" >> ${stop_file}
+if ${cygwin}; then
+    echo "echo \"Please manually observe the mongod processes, decide which ones apply\"" >> ${stop_file}
+    echo "echo \"to the replicaset in question, and stop them by hand.\"" >> ${stop_file}
+else
+    echo "ps ax | grep mongod | grep ${rsetid} | grep -v grep | sed 's/^[ \t]*//' | cut -d' ' -f1 | xargs kill -2" >> ${stop_file}
+fi
 
+#generate the replica set config script
 generate_rset_config
 
-echo "letting the mongo processes start before applying the replica set configuration..."
+#start the mongods and apply the replica set config script
+echo "starting the mongod processes..."
+sh ${start_file}
+wait `echo $!`
+#TODO get smarter as to how long to wait before trying to apply the replica set
+#     configuration by for example asking the mongod-s what state they are in.
+echo "letting the mongod processes start before applying the replica set configuration..."
 sleep ${sleep_secs} 
 mongo 127.0.0.1:50000 ${rsetid}.js
